requirements.txt
python-telegram-bot==21.6
google-auth==2.35.0
google-api-python-client==2.149.0
httpx==0.27.2
